requirements.txt
aws-cdk-lib>=2.96.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.27.93
aws-cdk.aws-apigatewayv2-alpha>=2.96.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.96.0a0
aws-cdk.aws_apigatewayv2_authorizers_alpha===2.96.0a0